[no-issue] fix: 채팅방 정보 조회 API 수정
